map/layers: extracted clipboard to own module.
