+++ src/renderer/map/Map.js
@@ -14,6 +14,7 @@ import layers from './layers'
 import basemap from './basemap'
 import './style/scalebar.css'
 import disposable from '../../shared/disposable'
+import clipboard from './clipboard'
 
 const zoom = view => view.getZoom()
 const center = view => toLonLat(view.getCenter())
@@ -108,8 +109,8 @@ const Map = props => {
   return <div
     id={props.id}
     tabIndex="0"
-    onFocus={() => evented.emit('MAP_FOCUS')}
-    onBlur={() => evented.emit('MAP_BLUR')}
+    onFocus={clipboard.engage}
+    onBlur={clipboard.disengage}
   />
 }
