+++ src/renderer/map/clipboard.js
@@ -1,0 +1,125 @@
+import { ipcRenderer } from 'electron'
+import Mousetrap from 'mousetrap'
+import { GeoJSON } from 'ol/format'
+
+import undo from '../undo'
+import evented from '../evented'
+import selection from '../selection'
+import inputLayers from '../project/layers'
+import Feature from '../project/Feature'
+import { noop } from '../../shared/combinators'
+
+/**
+ * features :: [ol/Feature]
+ * Current set of loaded features.
+ */
+const features = {}
+
+/**
+ * GeoJSON data, by definitions, comes in WGS84;
+ * OL uses Web-Mercator by default.
+ */
+const geoJSON = new GeoJSON({
+  dataProjection: 'EPSG:4326', // WGS84
+  featureProjection: 'EPSG:3857' // Web-Mercator
+})
+
+/**
+ * clipboardWrite :: [string] -> unit
+ * Write serialize (JSON) features to clipboard.
+ */
+const clipboardWrite = featureIds => {
+  const writeFeatures = feature => [Feature.layerId(feature), geoJSON.writeFeature(feature)]
+  const content = featureIds
+    .map(featureId => features[featureId])
+    .map(writeFeatures)
+
+  ipcRenderer.send('IPC_CLIPBOARD_WRITE', content)
+}
+
+/**
+ * editSelectAll :: () -> unit
+ */
+const editSelectAll = () => {
+  const featureIds = Object.values(features)
+    .filter(Feature.unlocked)
+    .filter(Feature.showing)
+    .map(Feature.id)
+
+  selection.deselect()
+  selection.select(featureIds)
+}
+
+/**
+ * editDelete :: () -> unit
+ * Delete selected features.
+ */
+const editDelete = () =>
+  inputLayers.removeFeatures(selection.selected('feature:'))
+
+/**
+ * editCut :: () -> unit
+ * Write current selection to clipboard and delete selected features.
+ */
+const editCut = () => {
+  const featureIds = selection.selected('feature:')
+  clipboardWrite(featureIds)
+  inputLayers.removeFeatures(featureIds)
+}
+
+/**
+ * editCopy :: () -> unit
+ * Write current selection to clipboard.
+ */
+const editCopy = () =>
+  clipboardWrite(selection.selected('feature:'))
+
+/**
+ * editPaste :: () -> unit
+ * Insert features from clipboard.
+ */
+const editPaste = async () => {
+  const content = await ipcRenderer.invoke('IPC_CLIPBOARD_READ')
+  if (!content) return
+  inputLayers.addFeatures(content)
+}
+
+
+Mousetrap.bind('del', editDelete) // macOS: fn+backspace
+Mousetrap.bind('command+backspace', editDelete)
+Mousetrap.bind('esc', () => selection.deselect())
+
+evented.on('EDIT_CUT', editCut)
+evented.on('EDIT_COPY', editCopy)
+evented.on('EDIT_PASTE', editPaste)
+
+
+// Only handle the following when map has focus.
+
+const engage = () => {
+  evented.on('EDIT_UNDO', undo.undo)
+  evented.on('EDIT_REDO', undo.redo)
+  evented.on('EDIT_SELECT_ALL', editSelectAll)
+}
+
+const disengage = () => {
+  evented.off('EDIT_UNDO', undo.undo)
+  evented.off('EDIT_REDO', undo.redo)
+  evented.off('EDIT_SELECT_ALL', editSelectAll)
+}
+
+const addFeature = feature => (features[Feature.id(feature)] = feature)
+const addFeatures = ({ features }) => features.forEach(addFeature)
+const deleteFeatures = ({ ids }) => ids.forEach(id => delete features[id])
+
+const eventHandlers = {
+  featuresadded: addFeatures,
+  featuresremoved: deleteFeatures
+}
+
+inputLayers.register(event => (eventHandlers[event.type] || noop)(event))
+
+export default {
+  engage,
+  disengage
+}
